+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [4.1.11] - 2017-07-02
 ### Added
 - Context `front_face` attribute
 
@@ -192,7 +194,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 1.0.0 - 2016-03-03
 - It was a tool for me to develop games in python
 
-[Unreleased]: https://github.com/cprogrammer1994/ModernGL/compare/4.1.10...master
+[Unreleased]: https://github.com/cprogrammer1994/ModernGL/compare/4.1.11...master
+[4.1.11]: https://github.com/cprogrammer1994/ModernGL/compare/4.1.10...4.1.11
 [4.1.10]: https://github.com/cprogrammer1994/ModernGL/compare/4.1.9...4.1.10
 [4.1.9]: https://github.com/cprogrammer1994/ModernGL/compare/4.1.8...4.1.9
 [4.1.8]: https://github.com/cprogrammer1994/ModernGL/compare/4.1.7...4.1.8